requirements.txt
pandas
matplotlib
seaborn
google-generativeai
reportlab
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-generativeai>=0.5.0
reportlab>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0